requirements.txt
backports.functools-lru-cache==1.6.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==2.2.4
nltk==3.4.5
numpy==1.16.5
oauthlib==3.1.0
pandas==0.24.2
Pillow==6.2.1
py4j==0.10.7
pyparsing==2.4.5
PySocks==1.7.1
pyspark==2.4.4
pyspark-stubs==2.4.0.post6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
singledispatch==3.4.0.3
six==1.13.0
subprocess32==3.5.4
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.7
wordcloud==1.6.0
